+++ Procfile
@@ -1,3 +1,3 @@
 web: /opt/venv/bin/gunicorn --config gunicorn.conf.py controller.webapp.wsgi
 service: /opt/venv/bin/python -m controller.service
-release: /opt/venv/bin/python -m jobrunner.cli.controller.migrate
+release: /opt/venv/bin/python -m controller.cli.migrate
