+++ controller/webapp/management/commands/flags.py
@@ -1,6 +1,6 @@
 from django.core.management.base import BaseCommand
 
-from jobrunner.cli.controller import flags as flags_cli
+from controller.cli import flags as flags_cli
 
 
 class Command(BaseCommand):
