--- jobrunner/cli/add_job.py
@@ -1,5 +1,0 @@
-from jobrunner.cli.controller import add_job
-
-
-if __name__ == "__main__":
-    add_job.run()
