--- jobrunner/cli/prepare_for_reboot.py
@@ -1,5 +1,0 @@
-from jobrunner.cli.controller import prepare_for_reboot
-
-
-if __name__ == "__main__":
-    prepare_for_reboot.run()
